troubleshooting for teammate dev server issues
